[refactor] backend physics implementation across several modules
